+++ recipes/chefdk_julia.rb
@@ -1,0 +1,5 @@
+if node['chefdk_bootstrap']['install_chefdk_julia']
+  chef_gem 'chefdk-julia'
+
+  # TODO: add the config to enable julia in knife.rb/config.rb
+end