+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Install 64-bit git 2.7
 * Install [Git Credential Manager for Windows](https://github.com/Microsoft/Git-Credential-Manager-for-Windows),
 which replaces the deprecated git-credential-winstore.
+* Add option to install chefdk-julia
 
 ## 1.4.1
 * Install Vagrant via community Vagrant cookbook